Full multiprocessing support After splitting the main run() function up into several smaller functions and removing all shared resources and files, full multiprocessing was implemented.
All filenames and directory names are now in the "directories" list, and
"talys_input" is no longer a class attribute. The reason for this is to
both ease transfer of information between the functions and to prevent
the multiple processes to interfer with each other. Appears to be
working smoothly.