Improve multiprocessing and add logging
Added a logging system whos verbosity and logging level
can be given in the terminal. It is recommended that all printing is
done through self.logger.<level>(string) to easily control what is
written.

Altough not done, the multiprocessing has been expanded to specifically
set how many CPUs to be used when combined with --multi_talys.