Clean and rewrite code
Have tried to seperate the code into two additional files in an attempt
to make it easier to read. Classes that read files of different format
are placed in input_reader.py while misc functions are placed in
tools.py

Have begun on rewriting some of the core code to make it more flexible.
It currently supports all of TALYS' keywords as well as making some
keywords depend on each other