Add documentation and a few features
Every function has been given a docstring explaing the parameters, what
is return and a general overview of the algorithm used. They
algorithmpart can be made more verbose at a future stage. All variables
and the every logical block of a function have been attempted to be
given a short description, explaining its purpose.

The scaffolding for MPI is completed, though any future readers can
easily make it working if they know MPI.

The technical challenge of iterating through the different names for the
directories was overcome, but the code is not by any means beautiful. It
works by reading a list of the format [{name:style}, {name:style} ...]
where "name" is a descriptor for the script and style is a formatstring
describing the name of the directory. This formatstring is formatted by
a custom formatter, StyleFormatter, that can understand keywords that go
atleast 2 deep, "{mass[element]}".

The argument --enable_pausing has been added to make it possible to
pause the script when needed. This feature is not enabled by default as
it requires a new process to wait for the keystroke 'p'. What it
actually does is that it orders the processing running self._run() to
wait for something to be put on a multiprocessing.Queue in the
product-loop, meaning there will be some delay from clicking pause to
the script actually pauses.