Change syntax of inputfile to support dependent dictionaries
If a keyword is given as a dict, the keys are assumed to be the element. The values
are "resolved" by inserting local variables into format strings. For example
```json
"E1file": {"Ce": "cefile.dat {Z} {A}", "Dy": "file2.dat {Z}"}
`
The strings will be resolved to "cefile.dat 58 114" and "file2.dat 66 160", with the former used for Ce and latter
 for Dy.
 Only tested for E1, but should work for other keywords aswell.